requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.0
webdriver-manager==4.0.1
schedule==1.2.0
configparser==6.0.0
python-telegram-bot==20.7
aiohttp==3.9.1
fastapi>=0.104.0
uvicorn>=0.24.0
